Namespace moved to HttpFoundation/Tests and use setCookie / new Cookie
