--- src/Symfony/Component/HttpFoundation/Tests/Session/Storage/Handler/Fixtures/native_header.expected
@@ -1,9 +1,0 @@
-
-Array
-(
-    [0] => Content-Type: text/plain; charset=utf-8
-    [1] => Cache-Control: no-cache, private
-    [2] => Date: Thu, 05 Apr 2018 12:12:12 GMT
-    [3] => Set-Cookie: key=value; path=/; domain=example.org; httponly; samesite=strict
-)
-shutdown
